// src/api/api.ts
export interface ActionDto {
    /**
     * Уникальный идентификатор действия
     * @type {string}
     * @memberof ActionDto
     */
    id: string;
    /**
     * Название действия
     * @type {string}
     * @memberof ActionDto
     */
    name: string;
    /**
     * Описание действия
     * @type {string}
     * @memberof ActionDto
     */
    description: string;
    /**
     * Стоимость действия
     * @type {number}
     * @memberof ActionDto
     */
    cost: number;
    /**
     * Прибыли и влияние действия (опционально)
     * @type {ProfitsDto}
     * @memberof ActionDto
     */
    profits?: ProfitsDto;
    /**
     * Список возможных событий для данного действия
     * @type {Array<GameEventDto>}
     * @memberof ActionDto
     */
    events: Array<GameEventDto>;
}
// ...
export const EventType = {
    BAD: 'BAD',
    GOOD: 'GOOD',
    SUPERBAD: 'SUPERBAD',
} as const;

export type EventType = (typeof EventType)[keyof typeof EventType];

/**
 *
 * @export
 * @interface GameDto
 */
export interface GameDto {
    /**
     * Начальный баланс игры
     * @type {number}
     * @memberof GameDto
     */
    statCash: number;
    /**
     * Начальный долг игры
     * @type {number}
     * @memberof GameDto
     */
    statTechDebt: number;
    /**
     * Начальная скорость разработки игры
     * @type {number}
     * @memberof GameDto
     */
    statDevelopingSpeed: number;
    /**
     * Список раундов в игре
     * @type {Array<RoundDto>}
     * @memberof GameDto
     */
    rounds: Array<RoundDto>;
}
/**
 *
 * @export
 * @interface GameEventDto
 */
export interface GameEventDto {
    /**
     * Уникальный идентификатор события
     * @type {string}
     * @memberof GameEventDto
     */
    id: string;
    /**
     * Тип события
     * @type {EventType}
     * @memberof GameEventDto
     */
    type: EventType;
    /**
     * Описание события
     * @type {string}
     * @memberof GameEventDto
     */
    description: string;
    /**
     * Прибыли и влияние проекта (опционально)
     * @type {ProfitsDto}
     * @memberof GameEventDto
     */
    profits?: ProfitsDto;
}
// ...
export interface ProfitsDto {
    /**
     * Влияние на скорость разработки
     * @type {number}
     * @memberof ProfitsDto
     */
    speed: number;
    /**
     * Влияние на технический долг
     * @type {number}
     * @memberof ProfitsDto
     */
    techDebt: number;
    /**
     * Влияние на зарплату/бюджет
     * @type {number}
     * @memberof ProfitsDto
     */
    cash: number;
}
/**
 *
 * @export
 * @interface ProjectDto
 */
export interface ProjectDto {
    /**
     * Уникальный идентификатор проекта
     * @type {string}
     * @memberof ProjectDto
     */
    id: string;
    /**
     * Название проекта
     * @type {string}
     * @memberof ProjectDto
     */
    name: string;
    /**
     * Описание проекта
     * @type {string}
     * @memberof ProjectDto
     */
    description: string;
    /**
     * Тип проекта
     * @type {string}
     * @memberof ProjectDto
     */
    type: ProjectDtoTypeEnum;
    /**
     * Стоимость проекта
     * @type {number}
     * @memberof ProjectDto
     */
    cost: number;
    /**
     * Прибыли и влияние проекта (опционально)
     * @type {ProfitsDto}
     * @memberof ProjectDto
     */
    profits?: ProfitsDto;
    /**
     * Список возможных событий для данного проекта
     * @type {Array<GameEventDto>}
     * @memberof ProjectDto
     */
    events: Array<GameEventDto>;
}

export const ProjectDtoTypeEnum = {
    product: 'product',
    marketing: 'marketing',
    techDebt: 'techDebt',
} as const;

export type ProjectDtoTypeEnum = (typeof ProjectDtoTypeEnum)[keyof typeof ProjectDtoTypeEnum];

/**
 *
 * @export
 * @interface QuizDto
 */
export interface QuizDto {
    /**
     * Уникальный идентификатор викторины
     * @type {string}
     * @memberof QuizDto
     */
    id: string;
    /**
     * Название категории викторины
     * @type {string}
     * @memberof QuizDto
     */
    name: string;
    /**
     * Описание категории викторины
     * @type {string}
     * @memberof QuizDto
     */
    description: string;
    /**
     * Тема викторины
     * @type {string}
     * @memberof QuizDto
     */
    theme: string;
    /**
     * Список юнитов/технологий в данной категории
     * @type {Array<UnitDto>}
     * @memberof QuizDto
     */
    units: Array<UnitDto>;
}
/**
 *
 * @export
 * @interface RoundDto
 */
export interface RoundDto {
    /**
     * Уникальный идентификатор раунда
     * @type {string}
     * @memberof RoundDto
     */
    id: string;
    /**
     * Название раунда
     * @type {string}
     * @memberof RoundDto
     */
    name: string;
    /**
     * Список викторин в раунде
     * @type {Array<QuizDto>}
     * @memberof RoundDto
     */
    quiz: Array<QuizDto>;
    /**
     * Список проектов в раунде
     * @type {Array<ProjectDto>}
     * @memberof RoundDto
     */
    projects: Array<ProjectDto>;
    /**
     * Список действий в раунде
     * @type {Array<ActionDto>}
     * @memberof RoundDto
     */
    actions: Array<ActionDto>;
}
// ...
export interface UnitDto {
    /**
     * Уникальный идентификатор юнита
     * @type {string}
     * @memberof UnitDto
     */
    id: string;
    /**
     * Название юнита/технологии
     * @type {string}
     * @memberof UnitDto
     */
    name: string;
    /**
     * Стоимость приобретения юнита
     * @type {number}
     * @memberof UnitDto
     */
    cost: number;
    /**
     * Прибыли и влияние юнита на различные метрики
     * @type {ProfitsDto}
     * @memberof UnitDto
     */
    profits: ProfitsDto;
    /**
     * Описание юнита
     * @type {string}
     * @memberof UnitDto
     */
    description: string;
    /**
     * Ссылка на иконку юнита
     * @type {string}
     * @memberof UnitDto
     */
    icon: string;
    /**
     * Список возможных событий для данного юнита
     * @type {Array<GameEventDto>}
     * @memberof UnitDto
     */
    events: Array<GameEventDto>;
}